Home

VisualDL.jl

This package provides a julia wrapper for VisualDL, which is a deep learning visualization tool that can help design deep learning jobs.

Currently, the wrapper is written on top of the Python SDK of VisualDL by PyCall. I have tried to write the wrapper on top of the C++ SDK by leveraging CxxWrap.jl. But unluckily a strange error encountered. Hopefully I'll figured it out later and swap the backend into C++.

Install

First, install the Python client of VisualDL. Checkout here for a detailed guide.

Then add this package as a dependent.

pkg> add https://github.com/findmyway/VisualDL.jl

Example

using VisualDL

# initialize
train_logger = VisualDLLogger("tmp", 1, "train")
test_logger = as_mode(train_logger, "test")

# log scalars
for i in 1:100
    with_logger(train_logger) do
        @log_scalar s0=(i,rand()) s1=(i, rand())
    end

    with_logger(test_logger) do
        @log_scalar s0=(i,rand()) s1=(i, rand())
    end
end

# log histograms
for i in 1:100
    with_logger(train_logger) do
       @log_histogram h0=(i, randn(100))
    end
end

# log texts
for i in 1:100
    with_logger(train_logger) do
       @log_text t0=(i, "This is test " * string(i))
    end
end

# log images
for i in 1:100
    with_logger(train_logger) do
       @log_image i0=([3,3,3], rand(27) * 255)
    end
end

for i in 1:100
    with_logger(test_logger) do
        # Array{Number, 3} is also supported
        @log_image image0=rand(10, 10, 3) * 255
    end
end

save(train_logger)
save(test_logger)

Reference

VisualDL.VisualDLLogger — Type.

VisualDLLogger is a subtype of AbstractLogger. And all the necessary interfaces are implemented. This means that you can use Base.CoreLogging.disable_logging to control the log level and use the similar grammar like @info.

The provided log levels are:

const ScalarLevel = LogLevel(880)
const HistogramLevel = LogLevel(881)
const EmbeddingLevel = LogLevel(882)
const TextLevel = LogLevel(883)
const AudioLevel = LogLevel(884)
const ImageLevel = LogLevel(885)

which are between the Info level(LogLevel(0)) andWarnlevel(LogLevel(1000)`).

There are two members in the VisualDLLogger, a pylogger which is a wrapper of the python logger instance, and a components which is a Dict containing all the components related to the logger.

source

VisualDL.VisualDLLogger — Type.

VisualDLLogger(log_path, sync_cycle, mode="train")

Create a logger instance. log_path is used to specify where to store the log data. sync_cycle specify how often should the system store data into the file system. Typically adding a record requires 6 operations. System will save the data into the file system once operations count reaches sync_cycle.

source

VisualDL.add_component — Method.

add_component(logger, level, tag; kwargs...)

Althrough this package will automatically initial a component for you when you add records to a component for the first time, it is strongly suggested to use this function to initial a component manually.

For the detail of kwargs, please read the python api

source

VisualDL.as_mode — Method.

as_mode(logger, mode)

Clone a logger and reset the mode.

source

VisualDL.finish_sampling — Method.

finish_sampling(logger, tag)

This function is only valid for @log_image and @log_audio.

source

VisualDL.save — Method.

save(logger)

Although the logger will automatically save data when the sync_cycle is reached. It is better to force save in the end.

source

VisualDL.set_caption — Method.

set_caption(logger, tag, caption)

Set the caption of a figure.

source

VisualDL.start_sampling — Method.

start_sampling(logger, tag)

This function is only valid for @log_image and @log_audio.

source

VisualDL.@log_histogram — Macro.

@log_scalar [tag_name=(args...) | ...]
@log_histogram [tag_name=(args...) | ...]
@log_embedding [tag_name=(args...) | ...]
@log_text [tag_name=(args...) | ...]
@log_audio [tag_name=(args...) | ...]
@log_image [tag_name=(args...) | ...]

Add a record to a component which has the tag_name. If the component is not initialized yet, a new component will be created automatically with default arguments. For the detail of args of each component, please read the doc

source

VisualDL.@log_image — Macro.

@log_scalar [tag_name=(args...) | ...]
@log_histogram [tag_name=(args...) | ...]
@log_embedding [tag_name=(args...) | ...]
@log_text [tag_name=(args...) | ...]
@log_audio [tag_name=(args...) | ...]
@log_image [tag_name=(args...) | ...]

Add a record to a component which has the tag_name. If the component is not initialized yet, a new component will be created automatically with default arguments. For the detail of args of each component, please read the doc

source

VisualDL.@log_scalar — Macro.

@log_scalar [tag_name=(args...) | ...]
@log_histogram [tag_name=(args...) | ...]
@log_embedding [tag_name=(args...) | ...]
@log_text [tag_name=(args...) | ...]
@log_audio [tag_name=(args...) | ...]
@log_image [tag_name=(args...) | ...]

Add a record to a component which has the tag_name. If the component is not initialized yet, a new component will be created automatically with default arguments. For the detail of args of each component, please read the doc

source

VisualDL.@log_text — Macro.

@log_scalar [tag_name=(args...) | ...]
@log_histogram [tag_name=(args...) | ...]
@log_embedding [tag_name=(args...) | ...]
@log_text [tag_name=(args...) | ...]
@log_audio [tag_name=(args...) | ...]
@log_image [tag_name=(args...) | ...]

Add a record to a component which has the tag_name. If the component is not initialized yet, a new component will be created automatically with default arguments. For the detail of args of each component, please read the doc

source

VisualDL.@log_audio — Macro.

@log_scalar [tag_name=(args...) | ...]
@log_histogram [tag_name=(args...) | ...]
@log_embedding [tag_name=(args...) | ...]
@log_text [tag_name=(args...) | ...]
@log_audio [tag_name=(args...) | ...]
@log_image [tag_name=(args...) | ...]

Add a record to a component which has the tag_name. If the component is not initialized yet, a new component will be created automatically with default arguments. For the detail of args of each component, please read the doc

source

VisualDL.@log_embedding — Macro.

@log_scalar [tag_name=(args...) | ...]
@log_histogram [tag_name=(args...) | ...]
@log_embedding [tag_name=(args...) | ...]
@log_text [tag_name=(args...) | ...]
@log_audio [tag_name=(args...) | ...]
@log_image [tag_name=(args...) | ...]

Add a record to a component which has the tag_name. If the component is not initialized yet, a new component will be created automatically with default arguments. For the detail of args of each component, please read the doc

source